shorts-to-analysis/requirements.txt
yt-dlp
opencv-python
youtube-transcript-api
google-generativeai
pandas
pillow
numpy
python-dotenv
scipy
requests
